feat: opt releases route
